+++ src/workflows/workflows.ts
@@ -20,7 +20,7 @@ import {
   registerOffchainWorkflowWithSigner,
 } from './registration';
 import { mintingWorkflow } from './minting';
-import { transfersWorkflow, batchTransfersWorkflow } from './transfers';
+import { transfersWorkflow, batchTransfersWorkflow, transfersWorkflowWithSigner } from './transfers';
 import {
   depositERC20Workflow,
   depositERC721Workflow,
@@ -113,12 +113,18 @@ export class Workflows {
     return mintingWorkflow(signer, request, this.mintsApi);
   }
 
+  /** @deprecated */
   public transfer(
     signer: Signer,
     starkWallet: StarkWallet,
     request: UnsignedTransferRequest,
   ) {
     return transfersWorkflow(signer, starkWallet, request, this.transfersApi);
+  }
+
+  public transferWithSigner(walletConnection: WalletConnection, request: UnsignedTransferRequest) {
+    return transfersWorkflowWithSigner(
+      { ...walletConnection, request, transfersApi: this.transfersApi  });
   }
 
   public batchNftTransfer(